+++ src/main/java/com/minhcraft/config/ModConfig.java
@@ -108,4 +108,7 @@ public class ModConfig extends MidnightConfig {
 
     @Entry(isSlider = true, min=0.0, max=1.0)
     public static double boatMaxSlipperiness = 0.9;
+
+    @Entry(min=0.0F, max=5.0)
+    public static float enderpearlTeleportDamage = 2.0F;
 }